Adding broken double mach-reflection.
Two things need to be sorted out in order for this application to work:

a) The boundary data for q to be sorted out in order for this application to
work with the Lax-Wendroff scheme.

b) The time needs to be saved in the "hackish" dogParams.set_time() in order
for this to work with an RK time-stepper.