Broken 2D apps now compile.
The culprit was the SetBndValuesX and SetBndValuesY that were added to
ConstructLFL_BCx2 inside of $FINESS/lib/2d/ConstructLFL.cpp.  The reason for
adding those calls was to accomodate problems with "geometry".  That is, for
problems such as the shock_diffraction problem, a total of two calls to
SetBndValues needed to be made: one for the X-direction, and one for the
Y-direction.