All 1D applications now compile.
Somethign was changed in the library Makefiles when references to
APP_LIB_SOURCES and APP_LIB_OBJECTS are defind.  When these variables are
empty, there are no problems, but when a user adds an extra file to these
calls, the corresponding Makefile (the one that actually gets called in the
application folder) cannot have the line

ALL_SOURCES = $(COMMON_SOURCES) $(APP_LIB_SOURCES)

instead it needs to have

ALL_SOURCES = $(COMMON_SOURCES)

Right now, the 2D/mhd as well as the 1d/isentropic_euler are written this way.
Every other application has the former.  I will have to invesitgate to see if
I want to adopt the latter.