Isent. Euler convergence example - multiple edits
1. Creating a new file, create_convergence_table.py that constructs a
convergence table from a list of error data.  This one asks for the following
two parameters: [mx, error] in a each row in a file.  (This is different than
previous versions, where the table creator required the L1, L2, and Li error
each time it was called.)

2. Fixed error when convergence factor is not 1.2.  The previous version of the
convergence analysis assumed that the refinement factor was always 1.2

3. Whitespace in $FINESS/apps/1d/isentropic_euler/lib/ConstructDiffTransformL.cpp